+++ src/biz/bokhorst/xprivacy/PackageChange.java
@@ -29,12 +29,16 @@ public class PackageChange extends BroadcastReceiver {
 
 				// Check action
 				if (intent.getAction().equals(Intent.ACTION_PACKAGE_ADDED)) {
+					// Check privacy service
+					if (PrivacyService.getClient() == null)
+						return;
+
 					// Get data
 					ApplicationInfoEx appInfo = new ApplicationInfoEx(context, uid);
 					String packageName = inputUri.getSchemeSpecificPart();
 
 					// Default deny new user apps
-					if (PrivacyService.getClient() != null && appInfo.getPackageName().size() == 1) {
+					if (appInfo.getPackageName().size() == 1) {
 						if (replacing)
 							PrivacyManager.clearPermissionCache(uid);
 						else {
@@ -44,16 +48,15 @@ public class PackageChange extends BroadcastReceiver {
 							PrivacyManager.deleteUsage(uid);
 							PrivacyManager.clearPermissionCache(uid);
 
-							// Enable on demand
-							boolean ondemand = PrivacyManager.getSettingBool(userId, PrivacyManager.cSettingOnDemand,
-									true);
-							if (ondemand)
-								PrivacyManager.setSetting(uid, PrivacyManager.cSettingOnDemand, Boolean.toString(true));
-
 							// Apply template
-							PrivacyManager.applyTemplate(uid, null, true, true);
+							PrivacyManager.applyTemplate(uid, Meta.cTypeTemplate, null, true, true);
 						}
 					}
+
+					// Enable on demand
+					boolean ondemand = PrivacyManager.getSettingBool(userId, PrivacyManager.cSettingOnDemand, true);
+					if (ondemand)
+						PrivacyManager.setSetting(uid, PrivacyManager.cSettingOnDemand, Boolean.toString(true));
 
 					// Mark as new/changed
 					PrivacyManager.setSetting(uid, PrivacyManager.cSettingState,
@@ -118,12 +121,16 @@ public class PackageChange extends BroadcastReceiver {
 					}
 
 				} else if (intent.getAction().equals(Intent.ACTION_PACKAGE_REMOVED) && !replacing) {
+					// Check privacy service
+					if (PrivacyService.getClient() == null)
+						return;
+
 					// Package removed
 					notificationManager.cancel(uid);
 
 					// Delete restrictions
 					ApplicationInfoEx appInfo = new ApplicationInfoEx(context, uid);
-					if (PrivacyService.getClient() != null && appInfo.getPackageName().size() == 0) {
+					if (appInfo.getPackageName().size() == 0) {
 						PrivacyManager.deleteRestrictions(uid, null, false);
 						PrivacyManager.deleteSettings(uid);
 						PrivacyManager.deleteUsage(uid);